Unify local storage and cookies across platforms
