+++ src/client/modules/user/helpers/LocalStorage.web.js
@@ -1,0 +1,9 @@
+export default class {
+  static async getItem(key) {
+    return window.localStorage.getItem(key);
+  }
+
+  static async setItem(key, value) {
+    window.localStorage.setItem(key, value);
+  }
+}
