+++ src/client/modules/user/index.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Expo from 'expo';
 import { Button } from 'react-native';
 import { SimpleLineIcons } from '@expo/vector-icons';
 import { StackNavigator } from 'react-navigation';
@@ -9,33 +8,9 @@ import { createTabBarIconWrapper } from '../common/components/native';
 import Profile from './containers/Profile';
 import Login from './containers/Login';
 import reducers from './reducers';
+import { tokenMiddleware, tokenAfterware, connectionParam } from './helpers/middleware';
 
 import Feature from '../connector';
-
-const tokenMiddleware = async (req, options, next) => {
-  options.headers['x-token'] = await Expo.SecureStore.getItemAsync('token');
-  options.headers['x-refresh-token'] = await Expo.SecureStore.getItemAsync('refreshToken');
-  next();
-};
-
-const tokenAfterware = async (res, options, next) => {
-  const token = options.headers['x-token'];
-  const refreshToken = options.headers['x-refresh-token'];
-  if (token) {
-    await Expo.SecureStore.setItemAsync('token', token);
-  }
-  if (refreshToken) {
-    await Expo.SecureStore.setItemAsync('refreshToken', refreshToken);
-  }
-  next();
-};
-
-const connectionParam = async () => {
-  return {
-    token: await Expo.SecureStore.getItemAsync('token'),
-    refreshToken: await Expo.SecureStore.getItemAsync('refreshToken')
-  };
-};
 
 class LoginScreen extends React.Component {
   static navigationOptions = ({ navigation }) => ({
