+++ src/client/modules/user/index.web.jsx
@@ -11,33 +11,9 @@ import ResetPassword from './containers/ResetPassword';
 import reducers from './reducers';
 
 import { AuthRoute, AuthNav, AuthLogin, AuthProfile } from './containers/Auth';
+import { tokenMiddleware, tokenAfterware, connectionParam } from './helpers/middleware';
 
 import Feature from '../connector';
-
-function tokenMiddleware(req, options, next) {
-  options.headers['x-token'] = window.localStorage.getItem('token');
-  options.headers['x-refresh-token'] = window.localStorage.getItem('refreshToken');
-  next();
-}
-
-function tokenAfterware(res, options, next) {
-  const token = options.headers['x-token'];
-  const refreshToken = options.headers['x-refresh-token'];
-  if (token) {
-    window.localStorage.setItem('token', token);
-  }
-  if (refreshToken) {
-    window.localStorage.setItem('refreshToken', refreshToken);
-  }
-  next();
-}
-
-function connectionParam() {
-  return {
-    token: window.localStorage.getItem('token'),
-    refreshToken: window.localStorage.getItem('refreshToken')
-  };
-}
 
 export default new Feature({
   route: [
